qt5base: fix nls support
Instead of hard coding the dependency to libiconv-full use the macros
provided by OpenWrt. In the gnu iconv library the second parameter of
iconv(), often called inbuf, is not const, but qt5 assumes that it is
in the gnu version of libiconv. If we make qt5 use the code for the
posix version of libiconv it does not add the -liconv link parameter.
This patch adds the -liconv link parameter.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>