qt5base: also build for host
Although Qt has the functionality of <host_build>'s,
it's supposed to be used and explicitly stated in
project files, so it can't be switched on from the
outside (e.g. via flags or env vars).

On top of that all builds using the <host_build>
directive get linked against libQtBootstrap.a which
is not always desired.

libQtBootstrap.a is also the only file available for
host builds, which means, <host_build>'s can't just make
use of and and link against e.g. libQtCore, as those
objects only get compiled for the target, not for the
host.

Because of above reasons, we build Qt twice now, once
for the host, once for the target.