qt6: add initial version for building and packaging base of Qt6
Qt6 is a largely complex beast and given the experience from previous
ports of Qt4 and Qt5, it will require quite some more work to accommodate
all needs and configurations.

Qt6 kind of ditched support for qmake in favour of cmake, at least when
it comes to building Qt6 and its official modules itself.

The qmake eco-system is and still needs to be supported though, as 3rd
party projects are still allowed to use qmake - for the time being at
least. qmake support for external projects untested and probably broken
at the moment, though.

Signed-off-by: Mirko Vogt <mirko-openwrt@nanl.de>