qt5quick2: Symlink $PATH/qmlplugindump to /bin/true
qmlplugindump appears to be called at the end of `make`-invocation for
qmake generated projects.

If it fails with exit code != 0 the OpenWrt build fails.

qmlplugindump is apparently needed as a host tool, but can not be
(easily) packaged as a host tool.

The project file states:

  "qmlplugindump cannot be a build tool, because it loads target plugins."

which seems pretty mysterious to me.

So it's needed as a requirement on the host system, and/but natively
installed (newer) versions - at with my Debian's version - started to
fail with:

  "QQmlComponent: Component is not ready"

I have no idea what qmlplugindump actually does or why it's needed.

It seems to do no harm to not invoke qmlplugindump - so
hack-not-even-fix it by just overriding each qmlplugindump invocation.